[feat]: InfraManager 추가 - 콘솔 애플리케이션 구동 역할
